requirements.txt
imageio==2.31.2
matplotlib==3.5.3
numpy==1.21.6
pandas==1.3.5
Pillow==9.2.0
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
torch==1.12.1
torchvision==0.13.1
tqdm==4.66.1
typing-extensions==4.3.0

